+++ custom_components/tuya_ble/devices.py
@@ -402,6 +402,16 @@ devices_database: dict[str, TuyaBLECategoryInfo] = {
                 manufacturer="Foxgard",
                 lock=1,
             ),
+            "kpn4zaf7": TuyaBLEProductInfo(
+                name="Invisible induction lock",
+                manufacturer="BSTUOKEY",
+                lock=1,
+            ),
+            "wgv4haro": TuyaBLEProductInfo(
+                name="Guard Dog Security Smart Lock",
+                manufacturer="Guard Dog Security",
+                lock=1,
+            ),
         },
     ),
     "dcb": TuyaBLECategoryInfo(
@@ -657,12 +667,15 @@ devices_database: dict[str, TuyaBLECategoryInfo] = {
     ),
     "ggq": TuyaBLECategoryInfo(
         products={
+            "jntxv3q4": TuyaBLEProductInfo(
+                name="YZD02B dual irrigation timer",
+            ),
             **dict.fromkeys(
                 ["6pahkcau", "hfgdqhho"],  # PPB A1  # SGW08  # device product_id
                 TuyaBLEProductInfo(
                     name="Irrigation computer",
                 ),
-            )
+            ),
         },
     ),
     "dd": TuyaBLECategoryInfo(