+++ custom_components/tuya_ble/switch.py
@@ -643,6 +643,7 @@ mapping: dict[str, TuyaBLECategorySwitchMapping] = {
                     "qycalacn",
                     "fnlw6npo",
                     "jjqi2syk",
+                    "jntxv3q4",
                 ],  # Irrigation computer - dual outlet
                 [
                     TuyaBLESwitchMapping(